requirements.txt
numpy 
pandas 
seaborn  
python-dotenv 
mysql-connector-python  
pymysql
scikit-learn
#-e .